+++ CHANGELOG.md
@@ -17,7 +17,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Module blacklist-filter functionality
 - Ask question to update module package files
 - Support for module source-directory and target-directory in composer.json
-
+    - `OxidEsales\ComposerPlugin\Installer\Package\ModulePackageInstaller::MODULES_DIRECTORY`
+    
 ## [5.1.1] - Unreleased
 
 ### Deprecated